+++ flake.nix
@@ -27,6 +27,10 @@
     packages = eachSystem (system: {
       inherit (pkgsFor.${system}) helix;
       yazelix_helix = pkgsFor.${system}.helix;
+      yazelix_helix_steel = pkgsFor.${system}.runCommand "yazelix-helix-steel" {} ''
+        install -D -m 0444 ${./yazelix/steel/bridge-actions.scm} \
+          "$out/share/yazelix-helix/steel/bridge-actions.scm"
+      '';
       /*
       The default Helix build. Uses the latest stable Rust toolchain, and unstable
       nixpkgs.